Move Epoch ddetail
