+++ src/pages/EpochDetail/index.tsx
